refactor: separate buttton click handler
